main.ts starts server
